+++ src/store/index.ts
@@ -1,10 +1,10 @@
 import { createStore } from 'vuex';
 import { RootStateInterface } from '@/store/types';
-import { authorizationStore } from '@/store/authorization';
+import { authorizationStore } from '@/user/store/authorization';
 import { cooperationStore } from '@/cooperation/store';
 import { localStorageStore } from '@/store/localstorage';
 import { housesStore } from '@/houses/store';
-import { userStore } from '@/store/user';
+import { userStore } from '@/user/store/user';
 import { pollsStore } from '@/store/polls';
 import { invitationsStore } from '@/store/invitations';
 import { apartmentsStore } from '@/apartment/store/apartments';
