+++ src/store/types.ts
@@ -1,6 +1,6 @@
 import { AxiosError, AxiosResponse } from 'axios';
-import { UserStateInterface } from '@/store/user/types';
-import { AuthorizationStateInterface } from '@/store/authorization/types';
+import { UserStateInterface } from '@/user/store/user/types';
+import { AuthorizationStateInterface } from '@/user/store/authorization/types';
 import { CooperationStateInterface } from '@/cooperation/store/types';
 import { LocalStorageStateInterface } from '@/store/localstorage/types';
 import { HousesStateInterface } from '@/houses/store/types';
