+++ src/user/models/update-user.model.ts
@@ -1,7 +1,7 @@
-import { UserDataInterface } from '@/store/user/types';
-import { UserContactInterface } from './../user/types';
-export class PostUserModel {
-  public registration_token: string;
+import { UpdateUserInterface } from '../store/authorization/types';
+import { UserContactInterface } from '../store/user/types';
+
+export class UpdateUserModel {
   public first_name: string;
   public middle_name: string;
   public last_name: string;
@@ -9,13 +9,12 @@ export class PostUserModel {
   public password: string;
   public contacts: UserContactInterface[];
 
-  constructor(userData: UserDataInterface) {
-    this.registration_token = userData.registrationToken;
+  constructor(userData: UpdateUserInterface) {
     this.first_name = userData.firstName;
     this.middle_name = userData.middleName;
     this.last_name = userData.lastName;
     this.email = userData.email;
-    this.password = userData.password;
+    this.password = 'passworD123';
     this.contacts = userData.contacts;
   }
 }
