+++ src/user/store/authorization/getters.ts
@@ -1,7 +1,7 @@
 import { GetterTree } from 'vuex';
 import { RootStateInterface } from '@/store/types';
-import { AuthorizationStateInterface, Getters } from '@/store/authorization/types';
-import { UserModel } from './models/user.model';
+import { AuthorizationStateInterface, Getters } from '@/user/store/authorization/types';
+import { UserModel } from '../../models/user.model';
 
 export const getters: GetterTree<AuthorizationStateInterface, RootStateInterface> & Getters = {
   loggedIn(state): boolean {
