+++ src/user/store/user/actions.ts
@@ -1,8 +1,14 @@
 import { ActionTree } from 'vuex';
 import { RootStateInterface } from '@/store/types';
-import { UserStateInterface, UserActionEnum, UserMutationEnum, Actions, UserDataInterface } from '@/store/user/types';
+import {
+  UserStateInterface,
+  UserActionEnum,
+  UserMutationEnum,
+  Actions,
+  UserDataInterface,
+} from '@/user/store/user/types';
 import { HTTP } from '@/core/api/http-common';
-import { PostUserModel } from '@/store/models/post-user.model';
+import { PostUserModel } from '@/user/models/post-user.model';
 
 export const actions: ActionTree<UserStateInterface, RootStateInterface> & Actions = {
   [UserActionEnum.SET_USER_INFO]: async ({ commit }, userData: UserDataInterface) => {
