+++ src/user/store/user/getters.ts
@@ -1,7 +1,7 @@
 import { GetterTree } from 'vuex';
 
 import { RootStateInterface } from '@/store/types';
-import { UserStateInterface, Getters } from '@/store/user/types';
+import { UserStateInterface, Getters } from '@/user/store/user/types';
 
 export const getters: GetterTree<UserStateInterface, RootStateInterface> & Getters = {
   getErrorMessage: (state) => {
